Refactor article feature to links feature
Move LinksTopView to links feature directory and remove article-related components and hooks
