+++ feature/links/components/cards/FeaturedCard.tsx
@@ -1,7 +1,7 @@
 import { Image, View } from "react-native";
 
 import { ThemedText } from "@/components/text/ThemedText";
-import { useOGData } from "@/feature/article/hooks/useOGData";
+import { useOGData } from "@/feature/links/hooks/useOGData";
 
 type FeaturedArticleCardProps = {
   full_url: string;
